ui: 🎨 move providers to own component
