+++ src/modules/swap/components/SwapProvidersList.vue
@@ -2,40 +2,6 @@
   <div class="modules--swap--components--swap-providers-list my-5">
     <!--
     =====================================================================================
-      Sceleton Loader (desktop/mobile)
-    =====================================================================================
-    -->
-    <div v-if="isLoading" class="my-5">
-      <div class="loading align-center px-5 py-3 rate d-none d-sm-flex">
-        <v-container fluid class="d-flex flex-column align-center my-3">
-          <img :src="currentPicture" height="30" class="my-4" />
-          <div class="titlePrimary--text font-weight-medium py-5">
-            Finding best rates...
-          </div>
-          <v-progress-linear indeterminate color="primary" />
-        </v-container>
-      </div>
-
-      <div
-        class="
-          loading
-          align-center
-          px-5
-          py-5
-          border-radius--10px
-          d-flex d-sm-none
-        "
-      >
-        <div class="titlePrimary--text font-weight-medium mew-body">
-          Finding best rates...
-        </div>
-        <v-spacer />
-        <v-skeleton-loader type="avatar" />
-      </div>
-    </div>
-
-    <!--
-    =====================================================================================
       Provider Rate Row
     =====================================================================================
     -->
@@ -170,7 +136,6 @@
 <script>
 import AppUserMsgBlock from '@/core/components/AppUserMsgBlock';
 import { formatFloatingPointValue } from '@/core/helpers/numberFormatHelper';
-import { _ } from 'web3-utils';
 const MAX_PROVIDERS = 3;
 export default {
   name: 'SwapProvidersList',
@@ -202,10 +167,6 @@ export default {
       type: Boolean,
       default: false
     },
-    checkLoading: {
-      type: Boolean,
-      default: true
-    },
     providersError: {
       type: Object,
       default: () => {
@@ -215,14 +176,7 @@ export default {
   },
   data() {
     return {
-      showMore: false,
-      currentPicture: null,
-      partners: [
-        { image: require('../assets/0x.png') },
-        { image: require('../assets/paraswap.png') },
-        { image: require('../assets/1inch.png') },
-        { image: require('../assets/Changelly.png') }
-      ]
+      showMore: false
     };
   },
   computed: {
@@ -312,55 +266,10 @@ export default {
           }
         });
       }
-    },
-    isLoading(value) {
-      if (value) {
-        let index = 0;
-        const showProviders = () => {
-          _.delay(
-            () => {
-              if (!this.checkLoading && index >= 4) {
-                this.$emit('stopLoadingProviders', false);
-              }
-
-              this.currentPicture = this.partners[index]?.image;
-
-              if (this.checkLoading && index >= 4) {
-                index = 0;
-              } else {
-                index++;
-              }
-
-              if (index <= this.partners.length) {
-                showProviders();
-              }
-            },
-            index === 0 ? 0 : 600
-          );
-        };
-        showProviders();
-      }
     }
   }
 };
 </script>
-
-<style lang="scss">
-.modules--swap--components--swap-providers-list {
-  .loading {
-    .v-skeleton-loader__avatar {
-      height: 25px !important;
-      width: 25px !important;
-    }
-    .v-skeleton-loader__heading {
-      height: 25px !important;
-      width: 100px !important;
-      border-radius: 15px !important;
-    }
-  }
-}
-</style>
-
 <style lang="scss" scoped>
 .rate-active {
   border: 1px solid var(--v-primary-base) !important;
